+++ internal/api/note.go
@@ -20,21 +20,32 @@ const (
 func FindAllNotes(s storage.Store) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		var err error
-		notes := []model.Note{}
+		noteList := []model.Note{}
 
 		fields := []string{"id", "created_at", "updated_at", "note"}
 		argsStr, argsInt := SetArgs(r, fields)
 
 		schema := r.Context().Value("schema").(string)
-		notes, err = s.Notes().FindAll(argsStr, argsInt, schema)
+		noteList, err = s.Notes().FindAll(argsStr, argsInt, schema)
 
 		if err != nil {
 			RespondWithError(w, http.StatusNotFound, err.Error())
 			return
 		}
 
-		notes = app.DecryptNotes(notes)
-		RespondWithJSON(w, http.StatusOK, notes)
+		// notes = app.DecryptNotes(notes)
+
+		// Encrypt payload
+		var payload model.Payload
+		key := r.Context().Value("transmissionKey").(string)
+		encrypted, err := app.EncryptJSON(key, noteList)
+		if err != nil {
+			RespondWithError(w, http.StatusInternalServerError, err.Error())
+			return
+		}
+		payload.Data = string(encrypted)
+
+		RespondWithJSON(w, http.StatusOK, payload)
 	}
 }
 
@@ -61,21 +72,45 @@ func FindNoteByID(s storage.Store) http.HandlerFunc {
 			return
 		}
 
-		RespondWithJSON(w, http.StatusOK, model.ToNoteDTO(uNote))
+		noteDTO := model.ToNoteDTO(uNote)
+
+		// Encrypt payload
+		var payload model.Payload
+		key := r.Context().Value("transmissionKey").(string)
+		encrypted, err := app.EncryptJSON(key, noteDTO)
+		if err != nil {
+			RespondWithError(w, http.StatusInternalServerError, err.Error())
+			return
+		}
+		payload.Data = string(encrypted)
+
+		RespondWithJSON(w, http.StatusOK, payload)
 	}
 }
 
 // Create ...
 func CreateNote(s storage.Store) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		var noteDTO model.NoteDTO
 
+		// TODO BEGIN: This part should be in a helper function
+		// Unmarshal request body to payload
+		var payload model.Payload
 		decoder := json.NewDecoder(r.Body)
-		if err := decoder.Decode(&noteDTO); err != nil {
+		if err := decoder.Decode(&payload); err != nil {
 			RespondWithError(w, http.StatusBadRequest, InvalidRequestPayload)
 			return
 		}
 		defer r.Body.Close()
+		// TODO END:
+
+		// Decrypt payload
+		var noteDTO model.NoteDTO
+		key := r.Context().Value("transmissionKey").(string)
+		err := app.DecryptJSON(key, []byte(payload.Data), &noteDTO)
+		if err != nil {
+			RespondWithError(w, http.StatusInternalServerError, err.Error())
+			return
+		}
 
 		schema := r.Context().Value("schema").(string)
 		createdNote, err := app.CreateNote(s, &noteDTO, schema)
@@ -84,7 +119,17 @@ func CreateNote(s storage.Store) http.HandlerFunc {
 			return
 		}
 
-		RespondWithJSON(w, http.StatusOK, model.ToNoteDTO(createdNote))
+		createdNoteDTO := model.ToNoteDTO(createdNote)
+
+		// Encrypt payload
+		encrypted, err := app.EncryptJSON(key, createdNoteDTO)
+		if err != nil {
+			RespondWithError(w, http.StatusInternalServerError, err.Error())
+			return
+		}
+		payload.Data = string(encrypted)
+
+		RespondWithJSON(w, http.StatusOK, payload)
 	}
 }
 
@@ -98,13 +143,23 @@ func UpdateNote(s storage.Store) http.HandlerFunc {
 			return
 		}
 
-		var noteDTO model.NoteDTO
+		// Unmarshal request body to payload
+		var payload model.Payload
 		decoder := json.NewDecoder(r.Body)
-		if err := decoder.Decode(&noteDTO); err != nil {
+		if err := decoder.Decode(&payload); err != nil {
 			RespondWithError(w, http.StatusBadRequest, InvalidRequestPayload)
 			return
 		}
 		defer r.Body.Close()
+
+		// Decrypt payload
+		var noteDTO model.NoteDTO
+		key := r.Context().Value("transmissionKey").(string)
+		err = app.DecryptJSON(key, []byte(payload.Data), &noteDTO)
+		if err != nil {
+			RespondWithError(w, http.StatusInternalServerError, err.Error())
+			return
+		}
 
 		schema := r.Context().Value("schema").(string)
 		note, err := s.Notes().FindByID(uint(id), schema)
@@ -119,7 +174,17 @@ func UpdateNote(s storage.Store) http.HandlerFunc {
 			return
 		}
 
-		RespondWithJSON(w, http.StatusOK, model.ToNoteDTO(updatedNote))
+		updatedNoteDTO := model.ToNoteDTO(updatedNote)
+
+		// Encrypt payload
+		encrypted, err := app.EncryptJSON(key, updatedNoteDTO)
+		if err != nil {
+			RespondWithError(w, http.StatusInternalServerError, err.Error())
+			return
+		}
+		payload.Data = string(encrypted)
+
+		RespondWithJSON(w, http.StatusOK, payload)
 	}
 }