+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - SendEmail function to app package
 - Email verification for new users
 - AES encrypt and decrypt functions
+- Encryption to all jsons in API
 ### Changed
 - Github organization name pass-wall to passwall